Fix crash on query of next persistent notification id
Using SetInteger prevents release check: 'Trying to set pref
persistent_notifications.next_id of type integer to value of type
string'. This also mimics chromium notification service code.

+++ src/core/platform_notification_service_qt.cpp
@@ -201,7 +201,7 @@ int64_t PlatformNotificationServiceQt::ReadNextPersistentNotificationId(content:
     Q_ASSERT(content::BrowserThread::CurrentlyOn(content::BrowserThread::UI));
     auto prefs = static_cast<ProfileQt *>(browser_context)->GetPrefs();
     int64_t nextId = prefs->GetInteger(prefs::kNotificationNextPersistentId) + 1;
-    prefs->SetInt64(prefs::kNotificationNextPersistentId, nextId);
+    prefs->SetInteger(prefs::kNotificationNextPersistentId, nextId);
     return nextId;
 }
